Commit 11.10
Let it be known that as of right now, login is not broken. In fact, I can no only login with the users defined in my database which is perfect. I don't know that I changed anything from the problems I was having yesterday, however, and so I hope that, as I assume, this issue was network related, that I do not need to be connected to the internet just to use this driver. Still. At this point, it works.

Added log4j to servlets, major builds
Updated log4j console, file appender settings
Updated database login - FUNCTIONAL